separate startup hook
